+++ CanvasPlusPlayground/Features/People/Models/User.swift
@@ -26,12 +26,12 @@ class User: Cacheable {
     var enrollmentRoles: [EnrollmentType] {
         Array(
             Set(
-                enrollments.compactMap { EnrollmentType(rawValue: $0.type) }
+                enrollments?.compactMap { EnrollmentType(rawValue: $0.type) } ?? []
             )
         )
         .sorted { $0.rawValue < $1.rawValue }
     }
-    var enrollments: [EnrollmentAPI]
+    var enrollments: [EnrollmentAPI]?
 
     // MARK: Custom
     var courseId: String?
@@ -47,7 +47,7 @@ class User: Cacheable {
         self.email = userAPI.email
         self.pronouns = userAPI.pronouns
         self.role = userAPI.role
-        self.enrollments = userAPI.enrollments ?? []
+        self.enrollments = userAPI.enrollments
         self.tag = ""
     }
 
@@ -60,10 +60,7 @@ class User: Cacheable {
         self.pronouns = other.pronouns ?? self.pronouns
         self.role = other.role ?? self.role
 
-        // this logic is wrong, we should make `enrollments` optional and check if it's nil in `other`. for now ill keep it.
-        if !other.enrollments.isEmpty {
-            self.enrollments = other.enrollments
-        }
+        self.enrollments = other.enrollments ?? self.enrollments
     }
 
     var hasAvatar: Bool {
